Repackage from org.researchobject to org.commonwl.viewer
