+++ src/main/java/org/commonwl/viewer/domain/CWLElement.java
@@ -1,5 +1,3 @@
-package org.researchobject;
-
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -19,17 +17,39 @@ package org.researchobject;
  * under the License.
  */
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+package org.commonwl.viewer.domain;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CwlViewerApplicationTests {
+/**
+ * Represents the input or output of a workflow/tool
+ */
+public class CWLElement {
 
-	@Test
-	public void contextLoads() {
-	}
+    private String label;
+    private String doc;
+    private String type;
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+    public String getDoc() {
+        return doc;
+    }
+
+    public void setDoc(String doc) {
+        this.doc = doc;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
 
 }
